Print list of elements in a list

+++ 0x17-doubly_linked_lists/0-print_dlistint.c
@@ -1,12 +1,12 @@
 #include "lists.h"
 /**
  * print_dlistint - a function tat prints nodes in a list
- * h: pointer to nodes in list
+ * @h: pointer to nodes in list
  * Return: number of nodes
  */
 size_t print_dlistint(const dlistint_t *h)
 {
-	size_t i = 0;
+	int i = 0;
 	
 	if (h == NULL)
 		return (i);
